MICNO-730: move manager to own module
